refactor(channels/discord): consolidate over-split modules
Reverse two over-eager splits where each file had only one consumer:

A. stream-buffer.ts + message-chunk.ts → outbound.ts
   Both were outbound's internal mechanics (only consumer); merging keeps
   "agent text → Discord send" in one file.

B. access-policy.ts + stop-command.ts → inbound.ts
   Both are pre-receive gates called only by dispatchInbound; they belong
   with the rest of inbound concerns.

Net: 15 → 11 source files. Test files split alongside (outbound now has
3 test files for distinct surfaces: outbound, outbound-stream, outbound-chunk).
