+++ src/channels/discord/index.ts
@@ -13,12 +13,10 @@ import type { Logger } from "../../logging/logger.js";
 import { loggers } from "../../logging/logger.js";
 import { getIsotopesHome } from "../../paths.js";
 import { DedupeCache } from "./dedupe.js";
-import { handleInbound } from "./inbound.js";
+import { handleInbound, passesAllowlist, maybeHandleStop } from "./inbound.js";
 import { resolveAgentId, resolveSessionKey } from "./routing.js";
 import { createDiscordCallbacks } from "./outbound.js";
 import { reactToMessage } from "./react.js";
-import { passesAllowlist } from "./access-policy.js";
-import { maybeHandleStop } from "./stop-command.js";
 import { resolveGroupPolicy, resolveToken, mapGuildsForReceive } from "./config.js";
 import { extractDiscordMetadata, formatInboundMeta } from "./message-metadata.js";
 import { ThreadBindingManager } from "./thread-binding.js";
